added a testIndex with preperation for modularization
